+++ lib/src/qr_code_scanner.dart
@@ -4,12 +4,16 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+import 'lifecycle_event_handler.dart';
 import 'qr_scanner_overlay_shape.dart';
 import 'types/barcode.dart';
 import 'types/barcode_format.dart';
 import 'types/camera.dart';
 import 'types/camera_exception.dart';
 import 'types/features.dart';
+import 'web/flutter_qr_stub.dart'
+// ignore: uri_does_not_exist
+    if (dart.library.html) 'web/flutter_qr_web.dart';
 
 typedef QRViewCreatedCallback = void Function(QRViewController);
 typedef PermissionSetCallback = void Function(QRViewController, bool);
@@ -18,23 +22,21 @@ typedef PermissionSetCallback = void Function(QRViewController, bool);
 /// and the barcode scanner gets displayed.
 class QRView extends StatefulWidget {
   const QRView({
-    @required Key key,
-    @required this.onQRViewCreated,
+    required Key key,
+    required this.onQRViewCreated,
     this.overlay,
     this.overlayMargin = EdgeInsets.zero,
     this.cameraFacing = CameraFacing.back,
     this.onPermissionSet,
-    this.formatsAllowed,
-  })  : assert(key != null),
-        assert(onQRViewCreated != null),
-        super(key: key);
+    this.formatsAllowed = const <BarcodeFormat>[],
+  }) : super(key: key);
 
   /// [onQRViewCreated] gets called when the view is created
   final QRViewCreatedCallback onQRViewCreated;
 
   /// Use [overlay] to provide an overlay for the view.
   /// This can be used to create a certain scan area.
-  final ShapeBorder overlay;
+  final QrScannerOverlayShape? overlay;
 
   /// Use [overlayMargin] to provide a margin to [overlay]
   final EdgeInsetsGeometry overlayMargin;
@@ -46,7 +48,7 @@ class QRView extends StatefulWidget {
   final CameraFacing cameraFacing;
 
   /// Calls the provided [onPermissionSet] callback when the permission is set.
-  final PermissionSetCallback onPermissionSet;
+  final PermissionSetCallback? onPermissionSet;
 
   /// Use [formatsAllowed] to specify which formats needs to be scanned.
   final List<BarcodeFormat> formatsAllowed;
@@ -56,7 +58,15 @@ class QRView extends StatefulWidget {
 }
 
 class _QRViewState extends State<QRView> {
-  var _channel;
+  late MethodChannel _channel;
+  late LifecycleEventHandler _observer;
+
+  @override
+  void initState() {
+    super.initState();
+    _observer = LifecycleEventHandler(resumeCallBack: updateDimensions);
+    WidgetsBinding.instance!.addObserver(_observer);
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -70,13 +80,20 @@ class _QRViewState extends State<QRView> {
     );
   }
 
+  @override
+  void dispose() {
+    super.dispose();
+    WidgetsBinding.instance!.removeObserver(_observer);
+  }
+
+  Future<void> updateDimensions() async {
+    await QRViewController.updateDimensions(
+        widget.key as GlobalKey<State<StatefulWidget>>, _channel,
+        overlay: widget.overlay);
+  }
+
   bool onNotification(notification) {
-    Future.microtask(() => {
-          QRViewController.updateDimensions(widget.key, _channel,
-              scanArea: widget.overlay != null
-                  ? (widget.overlay as QrScannerOverlayShape).cutOutSize
-                  : 0.0)
-        });
+    updateDimensions();
     return false;
   }
 
@@ -87,7 +104,7 @@ class _QRViewState extends State<QRView> {
         Container(
           padding: widget.overlayMargin,
           decoration: ShapeDecoration(
-            shape: widget.overlay,
+            shape: widget.overlay!,
           ),
         )
       ],
@@ -96,56 +113,59 @@ class _QRViewState extends State<QRView> {
 
   Widget _getPlatformQrView() {
     Widget _platformQrView;
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        _platformQrView = AndroidView(
-          viewType: 'net.touchcapture.qr.flutterqr/qrview',
-          onPlatformViewCreated: _onPlatformViewCreated,
-          creationParams:
-              _QrCameraSettings(cameraFacing: widget.cameraFacing).toMap(),
-          creationParamsCodec: StandardMessageCodec(),
-        );
-        break;
-      case TargetPlatform.iOS:
-        _platformQrView = UiKitView(
-          viewType: 'net.touchcapture.qr.flutterqr/qrview',
-          onPlatformViewCreated: _onPlatformViewCreated,
-          creationParams:
-              _QrCameraSettings(cameraFacing: widget.cameraFacing).toMap(),
-          creationParamsCodec: StandardMessageCodec(),
-        );
-        break;
-      default:
-        throw UnsupportedError(
-            "Trying to use the default webview implementation for $defaultTargetPlatform but there isn't a default one");
+    if (kIsWeb) {
+      _platformQrView = createWebQrView(
+        onPlatformViewCreated: widget.onQRViewCreated,
+        cameraFacing: widget.cameraFacing,
+      );
+    } else {
+      switch (defaultTargetPlatform) {
+        case TargetPlatform.android:
+          _platformQrView = AndroidView(
+            viewType: 'net.touchcapture.qr.flutterqr/qrview',
+            onPlatformViewCreated: _onPlatformViewCreated,
+            creationParams:
+                _QrCameraSettings(cameraFacing: widget.cameraFacing).toMap(),
+            creationParamsCodec: StandardMessageCodec(),
+          );
+          break;
+        case TargetPlatform.iOS:
+          _platformQrView = UiKitView(
+            viewType: 'net.touchcapture.qr.flutterqr/qrview',
+            onPlatformViewCreated: _onPlatformViewCreated,
+            creationParams:
+                _QrCameraSettings(cameraFacing: widget.cameraFacing).toMap(),
+            creationParamsCodec: StandardMessageCodec(),
+          );
+          break;
+        default:
+          throw UnsupportedError(
+              "Trying to use the default qrview implementation for $defaultTargetPlatform but there isn't a default one");
+      }
     }
     return _platformQrView;
   }
 
   void _onPlatformViewCreated(int id) {
-    // We pass the cutout size so that the scanner respects the scan area.
-    var cutOutSize = 0.0;
-    if (widget.overlay != null) {
-      cutOutSize = (widget.overlay as QrScannerOverlayShape).cutOutSize;
-    }
-
     _channel = MethodChannel('net.touchcapture.qr.flutterqr/qrview_$id');
 
     // Start scan after creation of the view
     final controller = QRViewController._(
-        _channel, widget.key, cutOutSize, widget.onPermissionSet)
-      .._startScan(widget.key, cutOutSize, widget.formatsAllowed);
+        _channel,
+        widget.key as GlobalKey<State<StatefulWidget>>?,
+        widget.onPermissionSet,
+        widget.cameraFacing)
+      .._startScan(widget.key as GlobalKey<State<StatefulWidget>>,
+          widget.overlay, widget.formatsAllowed);
 
     // Initialize the controller for controlling the QRView
-    if (widget.onQRViewCreated != null) {
-      widget.onQRViewCreated(controller);
-    }
+    widget.onQRViewCreated(controller);
   }
 }
 
 class _QrCameraSettings {
   _QrCameraSettings({
-    this.cameraFacing,
+    this.cameraFacing = CameraFacing.unknown,
   });
 
   final CameraFacing cameraFacing;
@@ -158,9 +178,10 @@ class _QrCameraSettings {
 }
 
 class QRViewController {
-  QRViewController._(MethodChannel channel, GlobalKey qrKey, double scanArea,
-      PermissionSetCallback onPermissionSet)
-      : _channel = channel {
+  QRViewController._(MethodChannel channel, GlobalKey? qrKey,
+      PermissionSetCallback? onPermissionSet, CameraFacing cameraFacing)
+      : _channel = channel,
+        _cameraFacing = cameraFacing {
     _channel.setMethodCallHandler((call) async {
       switch (call.method) {
         case 'onRecognizeQR':
@@ -169,9 +190,9 @@ class QRViewController {
             final code = args['code'] as String;
             final rawType = args['type'] as String;
             // Raw bytes are only supported by Android.
-            final rawBytes = args['rawBytes'] as List<int>;
+            final rawBytes = args['rawBytes'] as List<int>?;
             final format = BarcodeTypesExtension.fromString(rawType);
-            if (format != null) {
+            if (format != BarcodeFormat.unknown) {
               final barcode = Barcode(code, format, rawBytes);
               _scanUpdateController.sink.add(barcode);
             } else {
@@ -180,15 +201,10 @@ class QRViewController {
           }
           break;
         case 'onPermissionSet':
-          await getSystemFeatures(); // if we have no permission all features will not be avaible
-          if (call.arguments != null) {
-            if (call.arguments as bool) {
-              _hasPermissions = true;
-            } else {
-              _hasPermissions = false;
-            }
+          if (call.arguments != null && call.arguments is bool) {
+            _hasPermissions = call.arguments;
             if (onPermissionSet != null) {
-              onPermissionSet(this, call.arguments as bool);
+              onPermissionSet(this, _hasPermissions);
             }
           }
           break;
@@ -197,29 +213,33 @@ class QRViewController {
   }
 
   final MethodChannel _channel;
+  final CameraFacing _cameraFacing;
   final StreamController<Barcode> _scanUpdateController =
       StreamController<Barcode>();
 
   Stream<Barcode> get scannedDataStream => _scanUpdateController.stream;
 
-  SystemFeatures _features;
-  bool _hasPermissions;
-
-  SystemFeatures get systemFeatures => _features;
+  bool _hasPermissions = false;
   bool get hasPermissions => _hasPermissions;
 
   /// Starts the barcode scanner
-  Future<void> _startScan(GlobalKey key, double cutOutSize,
-      List<BarcodeFormat> barcodeFormats) async {
+  Future<void> _startScan(GlobalKey key, QrScannerOverlayShape? overlay,
+      List<BarcodeFormat>? barcodeFormats) async {
     // We need to update the dimension before the scan is started.
-    QRViewController.updateDimensions(key, _channel, scanArea: cutOutSize);
-    return _channel.invokeMethod(
-        'startScan', barcodeFormats?.map((e) => e.asInt())?.toList() ?? []);
+    try {
+      await QRViewController.updateDimensions(key, _channel, overlay: overlay);
+      return await _channel.invokeMethod(
+          'startScan', barcodeFormats?.map((e) => e.asInt()).toList() ?? []);
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
   }
 
   /// Gets information about which camera is active.
   Future<CameraFacing> getCameraInfo() async {
     try {
+      var cameraFacing = await _channel.invokeMethod('getCameraInfo') as int;
+      if (cameraFacing == -1) return _cameraFacing;
       return CameraFacing
           .values[await _channel.invokeMethod('getCameraInfo') as int];
     } on PlatformException catch (e) {
@@ -238,7 +258,7 @@ class QRViewController {
   }
 
   /// Get flashlight status
-  Future<bool> getFlashStatus() async {
+  Future<bool?> getFlashStatus() async {
     try {
       return await _channel.invokeMethod('getFlashInfo');
     } on PlatformException catch (e) {
@@ -249,16 +269,25 @@ class QRViewController {
   /// Toggles the flashlight between available modes
   Future<void> toggleFlash() async {
     try {
-      await _channel.invokeMethod('toggleFlash') as bool;
+      await _channel.invokeMethod('toggleFlash') as bool?;
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
   }
 
-  /// Pauses barcode scanning
+  /// Pauses the camera and barcode scanning
   Future<void> pauseCamera() async {
     try {
       await _channel.invokeMethod('pauseCamera');
+    } on PlatformException catch (e) {
+      throw CameraException(e.code, e.message);
+    }
+  }
+
+  /// Stops barcode scanning and the camera
+  Future<void> stopCamera() async {
+    try {
+      await _channel.invokeMethod('stopCamera');
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
@@ -278,27 +307,41 @@ class QRViewController {
     try {
       var features =
           await _channel.invokeMapMethod<String, dynamic>('getSystemFeatures');
-      return SystemFeatures.fromJson(features);
+      if (features != null) {
+        return SystemFeatures.fromJson(features);
+      }
+      throw CameraException('Error', 'Could not get system features');
     } on PlatformException catch (e) {
       throw CameraException(e.code, e.message);
     }
   }
 
-  /// Disposes the barcode stream.
+  /// Stops the camera and disposes the barcode stream.
   void dispose() {
+    if (defaultTargetPlatform == TargetPlatform.iOS) stopCamera();
     _scanUpdateController.close();
   }
 
   /// Updates the view dimensions for iOS.
-  static void updateDimensions(GlobalKey key, MethodChannel channel,
-      {double scanArea}) {
+  static Future<bool> updateDimensions(GlobalKey key, MethodChannel channel,
+      {QrScannerOverlayShape? overlay}) async {
     if (defaultTargetPlatform == TargetPlatform.iOS) {
-      final RenderBox renderBox = key.currentContext.findRenderObject();
-      channel.invokeMethod('setDimensions', {
-        'width': renderBox.size.width,
-        'height': renderBox.size.height,
-        'scanArea': scanArea ?? 0
-      });
+      // Add small delay to ensure the render box is loaded
+      await Future.delayed(Duration(milliseconds: 300));
+      if (key.currentContext == null) return false;
+      final renderBox = key.currentContext!.findRenderObject() as RenderBox;
+      try {
+        await channel.invokeMethod('setDimensions', {
+          'width': renderBox.size.width,
+          'height': renderBox.size.height,
+          'scanArea': overlay?.cutOutSize ?? 0,
+          'scanAreaOffset': overlay?.cutOutBottomOffset ?? 0
+        });
+        return true;
+      } on PlatformException catch (e) {
+        throw CameraException(e.code, e.message);
+      }
     }
+    return false;
   }
 }